chore: rename MainContent to PostsListUI, PostList to PostsList, PostComponent to PostPreview components
